refactor(model): decouple and reformat base code
