+++ src/main.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-require_relative 'controller/minesweeper'
-require_relative 'view/minesweeper_display'
+require_relative 'controller/game_controller'
+require_relative 'view/game_view'
 require_relative 'model/board'
 
-a = MinesweeperController.create_board(16, 43)
-p a
+board = Board.new(5, 4)
+view = View.new
+game_controller = GameController.new(board, view)
+game_controller.print_board
