+++ src/view/minesweeper_display.rb
@@ -3,4 +3,8 @@
 # models the views and user interactions
 class View
   def initialize; end
+
+  def print_board(board)
+    p board
+  end
 end
